finished chapter 3
