+++ src/main.cpp
@@ -4,6 +4,9 @@
 #include <fstream>
 #include <string>
 
+#include "./vec3/vec3.h"
+#include "./vec3/vec3_utils.cpp"
+
 std::string path("./out/");
 
 int main() {
@@ -15,12 +18,10 @@ int main() {
 
     fs << "P3\n" << nx << " " << ny << "\n255\n";
     for (int j(ny-1); j>=0; --j) for (int i(0); i<nx; ++i) {
-        float r(float(i) / float(nx));
-        float g(float(j) / float(ny));
-        float b(0.2);
-        int ir(255.99 * r);
-        int ig(255.99 * g);
-        int ib(255.99 * b);
+        vec3 col(float(i) / float(nx), float(j) / float(ny), 0.2);
+        int ir(255.99 * col[0]);
+        int ig(255.99 * col[1]);
+        int ib(255.99 * col[2]);
 
         fs << ir << " " << ig << " " << ib << "\n";
     }
