+++ src/vec3/vec3.h
@@ -4,6 +4,7 @@
 #include <math.h>
 #include <stdlib.h>
 #include <iostream>
+#include <fstream>
 
 class vec3 {
 private:
